requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
certifi==2018.1.18
chardet==3.0.4
cli-helpers==2.2.1
click==8.0.4
colorama==0.3.7
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==3.1.3
django-debug-toolbar==3.2.4
django-filter==2.4.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
djangorestframework==3.12.2
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==16.10.1
jsonfield==3.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
mycli==1.25.0
mysqlclient==2.0.3
netifaces==0.10.4
PAM==0.4.2
prompt-toolkit==3.0.29
pyaes==1.6.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
Pygments==2.12.0
PyGObject==3.26.1
PyMySQL==1.0.2
pyOpenSSL==17.5.0
pyperclip==1.8.2
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
pytz==2022.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.11.0
sqlparse==0.4.2
ssh-import-id==5.7
swapper==1.3.0
systemd-python==234
tabulate==0.8.9
Twisted==17.9.0
typing_extensions==4.1.1
ufw==0.36
urllib3==1.22
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.6.0
zope.interface==4.3.2
